Move PACT tests / run separately / reduce test logging
