+++ logger.ts
@@ -3,6 +3,10 @@ import bunyanFormat from 'bunyan-format'
 
 const formatOut = bunyanFormat({ outputMode: 'short', color: true })
 
-const logger = bunyan.createLogger({ name: 'Manage Recalls', stream: formatOut, level: 'debug' })
+const logger = bunyan.createLogger({
+  name: 'Manage Recalls',
+  stream: formatOut,
+  level: process.env.NODE_ENV === 'test' ? 'fatal' : 'debug',
+})
 
 export default logger
