+++ pact/getDossier.pact.test.ts
@@ -1,9 +1,9 @@
 // @ts-nocheck
 import { pactWith } from 'jest-pact'
 import { Matchers } from '@pact-foundation/pact'
-import { getGeneratedDocument } from './manageRecallsApiClient'
-import * as configModule from '../../config'
-import getDossierResponseJson from '../../../fake-manage-recalls-api/stubs/__files/get-dossier.json'
+import { getGeneratedDocument } from '../server/clients/manageRecallsApi/manageRecallsApiClient'
+import * as configModule from '../server/config'
+import getDossierResponseJson from '../fake-manage-recalls-api/stubs/__files/get-dossier.json'
 import { pactGetRequest, pactJsonResponse } from './pactTestUtils'
 
 pactWith({ consumer: 'manage-recalls-ui', provider: 'manage-recalls-api' }, provider => {
