+++ pact/jest.pact.config.js
@@ -1,7 +1,5 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  verbose: true,
-  rootDir: '../',
-  testMatch: ['**/*.pact.test.(ts|js)'],
+  testMatch: ['**/*.pact.test.ts'],
 }
