+++ pact/searchPersons.pact.test.ts
@@ -1,10 +1,10 @@
 // @ts-nocheck
 import { pactWith } from 'jest-pact'
 import { Matchers } from '@pact-foundation/pact'
-import { createRecall } from './manageRecallsApiClient'
-import * as configModule from '../../config'
-import createRecallResponseJson from '../../../fake-manage-recalls-api/stubs/__files/create-recall.json'
-import { pactJsonResponse, pactPostRequest } from './pactTestUtils'
+import { searchByNomsNumber } from '../server/clients/manageRecallsApi/manageRecallsApiClient'
+import * as configModule from '../server/config'
+import searchResponseJson from '../fake-manage-recalls-api/stubs/__files/search.json'
+import { pactPostRequest, pactJsonResponse } from './pactTestUtils'
 
 pactWith({ consumer: 'manage-recalls-ui', provider: 'manage-recalls-api' }, provider => {
   const accessToken = 'accessToken-1'
@@ -14,17 +14,17 @@ pactWith({ consumer: 'manage-recalls-ui', provider: 'manage-recalls-api' }, prov
     jest.spyOn(configModule, 'manageRecallsApiConfig').mockReturnValue({ url: provider.mockService.baseUrl })
   })
 
-  describe('create a recall', () => {
-    test('can successfully create a recall', async () => {
+  describe('search prisoners', () => {
+    test('can find a prisoner by NOMS number', async () => {
       await provider.addInteraction({
-        state: 'a recall can be created',
-        ...pactPostRequest('a create recall request', '/recalls', { nomsNumber }, accessToken),
-        willRespondWith: pactJsonResponse(Matchers.like(createRecallResponseJson), 201),
+        state: `a prisoner exists for NOMS number`,
+        ...pactPostRequest('a search request by NOMS number', '/search', { nomsNumber }, accessToken),
+        willRespondWith: pactJsonResponse(searchResponseJson, 200),
       })
 
-      const actual = await createRecall(nomsNumber, accessToken)
+      const actualResults = await searchByNomsNumber(nomsNumber, accessToken)
 
-      expect(actual).toEqual(createRecallResponseJson)
+      expect(actualResults).toStrictEqual(searchResponseJson[0])
     })
 
     test('returns 400 if blank NOMS number provided', async () => {
@@ -34,10 +34,10 @@ pactWith({ consumer: 'manage-recalls-ui', provider: 'manage-recalls-api' }, prov
         message: 'nomsNumber: must not be blank',
       }
       await provider.addInteraction({
-        state: 'a recall can be created',
+        state: 'a search by blank NOMS number',
         ...pactPostRequest(
-          'a create recall request with blank nomsNumber',
-          '/recalls',
+          'a search request with blank NOMS number',
+          '/search',
           { nomsNumber: blankNomsNumber },
           accessToken
         ),
@@ -45,7 +45,7 @@ pactWith({ consumer: 'manage-recalls-ui', provider: 'manage-recalls-api' }, prov
       })
 
       try {
-        await createRecall(blankNomsNumber, accessToken)
+        await searchByNomsNumber(blankNomsNumber, accessToken)
       } catch (exception) {
         expect(exception.status).toEqual(400)
         expect(exception.data).toEqual(errorResponse)
@@ -55,12 +55,12 @@ pactWith({ consumer: 'manage-recalls-ui', provider: 'manage-recalls-api' }, prov
     test('returns 401 if invalid user', async () => {
       await provider.addInteraction({
         state: 'an unauthorized user accessToken',
-        ...pactPostRequest('an unauthorized create recall request', '/recalls', { nomsNumber }, accessToken),
+        ...pactPostRequest('an unauthorized search request', '/search', { nomsNumber }, accessToken),
         willRespondWith: { status: 401 },
       })
 
       try {
-        await createRecall(nomsNumber, accessToken)
+        await searchByNomsNumber(nomsNumber, accessToken)
       } catch (exception) {
         expect(exception.status).toEqual(401)
       }
