+++ pact/updateRecall.pact.test.ts
@@ -1,11 +1,11 @@
 // @ts-nocheck
 import { pactWith } from 'jest-pact'
 import { Matchers } from '@pact-foundation/pact'
-import { updateRecall } from './manageRecallsApiClient'
-import * as configModule from '../../config'
-import fullyPopulatedRecallJson from '../../../fake-manage-recalls-api/stubs/__files/get-recall.json'
-import recallWithSentencingInfoJson from '../../../fake-manage-recalls-api/stubs/__files/get-recall-with-sentencing-info.json'
-import emptyRecallJson from '../../../fake-manage-recalls-api/stubs/__files/get-recall-empty.json'
+import { updateRecall } from '../server/clients/manageRecallsApi/manageRecallsApiClient'
+import * as configModule from '../server/config'
+import fullyPopulatedRecallJson from '../fake-manage-recalls-api/stubs/__files/get-recall.json'
+import recallWithSentencingInfoJson from '../fake-manage-recalls-api/stubs/__files/get-recall-with-sentencing-info.json'
+import emptyRecallJson from '../fake-manage-recalls-api/stubs/__files/get-recall-empty.json'
 import { pactPatchRequest, pactJsonResponse } from './pactTestUtils'
 
 pactWith({ consumer: 'manage-recalls-ui', provider: 'manage-recalls-api' }, provider => {
